requirements.txt
# LangGraph and LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.17

# Environment management
python-dotenv==1.0.0

# Data science basics
pandas==2.2.0
numpy==1.26.3

# Utilities
pydantic==2.10.3

# Graph visualization
grandalf==0.8
